+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
   * `http.auth` module. Support for Basic auth
+  * `params` argument for the `HTTPClient.request` to simplify query parameters usage
 
 ## [0.1.2] - 2024-09-20
 ### Added